backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
aiosqlite==0.21.0
alembic==1.16.5
pydantic==2.11.9
pydantic-settings==2.11.0
python-jose[cryptography]==3.5.0
httpx==0.28.1
openai==1.107.1
orjson==3.11.3
